requirements.txt
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.4.4
bottle==0.12.18
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
conan==1.29.1
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.6.0
deprecation==2.0.7
distlib==0.3.1
distro==1.5.0
Django==2.1.7
django-allauth==0.39.1
django-cors-middleware==1.3.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extensions==2.1.5
django-rest-knox==4.1.0
djangorestframework==3.9.1
djangorestframework-api-key==2.0.0
fasteners==0.15
filelock==3.0.12
future==0.18.2
geographiclib==1.50
geopy==2.0.0
idna==2.10
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
node-semver==0.6.1
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
patch-ng==1.17.4
pep8==1.7.1
Pillow==6.2.2
pipenv==2020.8.13
pluggy==0.13.1
pluginbase==1.0.0
psycopg2==2.8.6
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
python-dateutil==2.8.1
python-decouple==3.1
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.1
stripe==2.27.0
tqdm==4.49.0
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
